+++ rust/crd/src/lib.rs
@@ -7,6 +7,7 @@ use kube::CustomResource;
 use schemars::JsonSchema;
 use semver::Version;
 use serde::{Deserialize, Serialize};
+use stackable_operator::identity::PodToNodeMapping;
 use stackable_operator::product_config_utils::{ConfigError, Configuration};
 use stackable_operator::role_utils::{CommonConfiguration, Role};
 use stackable_operator::status::{Conditions, Status, Versioned};
@@ -389,6 +390,8 @@ pub struct SparkClusterStatus {
     pub current_command: Option<CurrentCommand>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub cluster_execution_status: Option<ClusterExecutionStatus>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub history: Option<PodToNodeMapping>,
 }
 
 impl Versioned<SparkVersion> for SparkClusterStatus {