+++ rust/operator/src/lib.rs
@@ -1,8 +1,4 @@
-use std::collections::{BTreeMap, HashMap};
-use std::future::Future;
-use std::pin::Pin;
-use std::sync::Arc;
-use std::time::Duration;
+mod error;
 
 use async_trait::async_trait;
 use k8s_openapi::api::core::v1::{ConfigMap, EnvVar, Pod};
@@ -32,7 +28,12 @@ use stackable_operator::role_utils::{
     find_nodes_that_fit_selectors, get_role_and_group_labels,
     list_eligible_nodes_for_role_and_group, EligibleNodesForRoleAndGroup,
 };
-use stackable_operator::{configmap, k8s_utils, name_utils};
+use stackable_operator::{configmap, name_utils};
+use std::collections::{BTreeMap, HashMap};
+use std::future::Future;
+use std::pin::Pin;
+use std::sync::Arc;
+use std::time::Duration;
 use strum::IntoEnumIterator;
 use tracing::{debug, error, info, trace, warn};
 
@@ -46,16 +47,20 @@ use crate::config::validated_product_config;
 use crate::pod_utils::{
     filter_pods_for_type, get_hashed_master_urls, APP_NAME, MANAGED_BY, MASTER_URLS_HASH_LABEL,
 };
+use stackable_operator::identity::{LabeledPodIdentityFactory, PodIdentity, PodToNodeMapping};
+use stackable_operator::scheduler::{
+    K8SUnboundedHistory, RoleGroupEligibleNodes, ScheduleStrategy, Scheduler, StickyScheduler,
+};
 use stackable_operator::status::init_status;
 use stackable_operator::versioning::{finalize_versioning, init_versioning};
 
 mod command_utils;
 pub mod config;
-mod error;
 pub mod pod_utils;
 
 const FINALIZER_NAME: &str = "spark.stackable.tech/cleanup";
 const SHOULD_BE_SCRAPED: &str = "monitoring.stackable.tech/should_be_scraped";
+const ID_LABEL: &str = "spark.stackable.tech/id";
 
 const CONFIG_MAP_TYPE_CONF: &str = "config";
 
@@ -170,9 +175,8 @@ impl SparkState {
     }
 
     async fn create_missing_pods(&mut self) -> SparkReconcileResult {
-        trace!("Starting `create_missing_pods`");
+        trace!(target: "create_missing_pods","Starting `create_missing_pods`");
 
-        let mut changes_applied = false;
         // The iteration happens in two stages here, to accommodate the way our operators think
         // about roles and role groups.
         // The hierarchy is:
@@ -182,11 +186,11 @@ impl SparkState {
             let role_str = &role.to_string();
             if let Some(nodes_for_role) = self.eligible_nodes.get(role_str) {
                 for (role_group, eligible_nodes) in nodes_for_role {
-                    debug!(
+                    debug!( target: "create_missing_pods",
                         "Identify missing pods for [{}] role and group [{}]",
                         role_str, role_group
                     );
-                    trace!(
+                    trace!( target: "create_missing_pods",
                         "candidate_nodes[{}]: [{:?}]",
                         eligible_nodes.nodes.len(),
                         eligible_nodes
@@ -195,7 +199,7 @@ impl SparkState {
                             .map(|node| node.metadata.name.as_ref().unwrap())
                             .collect::<Vec<_>>()
                     );
-                    trace!(
+                    trace!(target: "create_missing_pods",
                         "existing_pods[{}]: [{:?}]",
                         &self.existing_pods.len(),
                         &self
@@ -204,62 +208,78 @@ impl SparkState {
                             .map(|pod| pod.metadata.name.as_ref().unwrap())
                             .collect::<Vec<_>>()
                     );
-                    trace!(
+                    trace!(target: "create_missing_pods",
                         "labels: [{:?}]",
                         get_role_and_group_labels(role_str, role_group)
                     );
-                    let nodes_that_need_pods = k8s_utils::find_nodes_that_need_pods(
-                        &eligible_nodes.nodes,
-                        &self.existing_pods,
-                        &get_role_and_group_labels(role_str, role_group),
-                        eligible_nodes.replicas,
+                    let mut history = match self
+                        .context
+                        .resource
+                        .status
+                        .as_ref()
+                        .and_then(|status| status.history.as_ref())
+                    {
+                        Some(simple_history) => {
+                            // we clone here because we cannot access mut self because we need it later
+                            // to create config maps and pods. The `status` history will be out of sync
+                            // with the cloned `simple_history` until the next reconcile.
+                            // The `status` history should not be used after this method to avoid side
+                            // effects.
+                            K8SUnboundedHistory::new(&self.context.client, simple_history.clone())
+                        }
+                        None => K8SUnboundedHistory::new(
+                            &self.context.client,
+                            PodToNodeMapping::default(),
+                        ),
+                    };
+
+                    let mut sticky_scheduler =
+                        StickyScheduler::new(&mut history, ScheduleStrategy::GroupAntiAffinity);
+
+                    let pod_id_factory = LabeledPodIdentityFactory::new(
+                        APP_NAME,
+                        &self.context.name(),
+                        &self.eligible_nodes,
+                        ID_LABEL,
+                        1,
                     );
 
-                    for node in nodes_that_need_pods {
-                        let node_name = if let Some(node_name) = &node.metadata.name {
-                            node_name
-                        } else {
-                            warn!("No name found in metadata, this should not happen! Skipping node: [{:?}]", node);
-                            continue;
-                        };
-                        debug!(
-                            "Creating pod on node [{}] for [{}] role and group [{}]",
-                            node.metadata
-                                .name
-                                .as_deref()
-                                .unwrap_or("<no node name found>"),
-                            role_str,
-                            role_group
-                        );
+                    let state = sticky_scheduler.schedule(
+                        &pod_id_factory,
+                        &RoleGroupEligibleNodes::from(&self.eligible_nodes),
+                        &self.existing_pods,
+                    )?;
 
-                        // now we have a node that needs pods -> get validated config
+                    let mapping = state.remaining_mapping().filter(
+                        APP_NAME,
+                        &self.context.name(),
+                        role_str,
+                        role_group,
+                    );
+
+                    if let Some((pod_id, node_id)) = mapping.iter().next() {
+                        // now we have a node that needs a pod -> get validated config
                         let validated_config = config_for_role_and_group(
-                            role_str,
-                            role_group,
+                            pod_id.role(),
+                            pod_id.group(),
                             &self.validated_role_config,
                         )?;
 
                         let config_maps = self
-                            .create_config_maps(role_str, role_group, validated_config)
+                            .create_config_maps(pod_id, validated_config, &state.mapping())
                             .await?;
 
                         self.create_pod(
+                            pod_id,
                             &role,
-                            role_group,
-                            node_name,
+                            &node_id.name,
                             &config_maps,
                             validated_config,
                         )
                         .await?;
 
-                        changes_applied = true;
-                    }
+                        history.save(&self.context.resource).await?;
 
-                    // we do this here to make sure pods of each role are started after each other
-                    // roles start in rolling fashion
-                    // pods of each role group start in parallel
-                    // master > worker > history server
-                    if changes_applied {
                         return Ok(ReconcileFunctionAction::Requeue(Duration::from_secs(10)));
                     }
                 }
@@ -293,19 +313,19 @@ impl SparkState {
     ///
     async fn create_config_maps(
         &self,
-        role: &str,
-        group: &str,
+        pod_id: &PodIdentity,
         validated_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
+        _pod_mapping: &PodToNodeMapping,
     ) -> Result<HashMap<&'static str, ConfigMap>, error::Error> {
         let mut config_maps = HashMap::new();
         let mut cm_conf_data = BTreeMap::new();
 
         let mut cm_labels = get_recommended_labels(
             &self.context.resource,
-            APP_NAME,
+            pod_id.app(),
             &self.context.resource.spec.version.to_string(),
-            role,
-            group,
+            pod_id.role(),
+            pod_id.group(),
         );
 
         cm_labels.insert(
@@ -350,10 +370,10 @@ impl SparkState {
         }
 
         let cm_conf_name = name_utils::build_resource_name(
-            APP_NAME,
+            pod_id.app(),
             &self.context.name(),
-            role,
-            Some(group),
+            pod_id.role(),
+            Some(pod_id.group()),
             None,
             Some(CONFIG_MAP_TYPE_CONF),
         )?;
@@ -386,8 +406,8 @@ impl SparkState {
     ///
     async fn create_pod(
         &self,
+        pod_id: &PodIdentity,
         role: &SparkRole,
-        group: &str,
         node_name: &str,
         config_maps: &HashMap<&'static str, ConfigMap>,
         validated_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
@@ -395,7 +415,6 @@ impl SparkState {
         let mut env_vars = vec![];
         let mut cli_arguments = vec![];
         let mut container_ports = (None, None);
-        let role_str = &role.to_string();
 
         // extract container ports
         if let Some(config) =
@@ -446,10 +465,10 @@ impl SparkState {
         }
 
         let pod_name = name_utils::build_resource_name(
-            APP_NAME,
+            pod_id.app(),
             &self.context.name(),
-            role_str,
-            Some(group),
+            pod_id.role(),
+            Some(pod_id.group()),
             Some(node_name),
             None,
         )?;
@@ -457,11 +476,12 @@ impl SparkState {
         let version = &self.context.resource.spec.version;
         let mut recommended_labels = get_recommended_labels(
             &self.context.resource,
-            APP_NAME,
+            pod_id.app(),
             &version.to_string(),
-            role_str,
-            group,
+            pod_id.role(),
+            pod_id.group(),
         );
+        recommended_labels.insert(ID_LABEL.to_string(), pod_id.id().to_string());
 
         let master_pods = filter_pods_for_type(&self.existing_pods, &SparkRole::Master);
         let master_urls = &config::get_master_urls(&master_pods, &self.validated_role_config)?;
@@ -549,6 +569,7 @@ impl SparkState {
             .node_name(node_name)
             .build()?;
 
+        trace!("create_pod: {:?}", pod_id);
         Ok(self.context.client.create(&pod).await?)
     }